Security-Operations-Agent/pyproject.toml
[project]
name = "cert-siem-poc-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "boto3>=1.42.13",
    "chromadb>=1.3.7",
    "dotenv>=0.9.9",
    "gradio>=6.1.0",
    "ollama>=0.6.1",
    "openai-agents[litellm]>=0.6.2",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "unstructured[all-docs]>=0.18.21",
]
